fatima commit description chek kerlo
ive added a scene Game 1 usmmain apnee level 2 ki changes kerdo, and use that as level 2. Moreover, ive added a Tank enemy prefab, use that. whenever u make a new enemy make a prefab of it by copy pasting the original enemy in the prefab folder and changing it.

Consequently, whenever u add an enemy to the game, make sure to drag and drop it in the array of GameManager gameobj.